requirements.txt
aiohttp==3.6.0
appnope==0.1.0
argcomplete==1.12.0
asn1crypto==1.4.0
async-exit-stack==1.0.1
async-generator==1.10
async-lru==0.1.0
async-service==0.1.0a8
async-timeout==3.0.1
asyncio-cancel-token==0.2.0
asyncio-run-in-process==0.1.0a8
attrs==20.1.0
backcall==0.2.0
base58==1.0.3
bitarray==1.2.2
blake2b-py==0.1.3
bloom-filter==1.3
boto3==1.14.54
botocore==1.17.54
cache==1.0.3
cached-property==1.5.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
coincurve==10.0.0
cryptography==2.8
cytoolz==0.10.1
decorator==4.4.2
docutils==0.15.2
eth-abi==2.1.1
eth-account==0.5.3
eth-bloom==1.0.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
fastapi==0.61.1
fastecdsa==1.7.5
Flask==1.1.2
Flask-WTF==0.14.3
h11==0.9.0
hexbytes==0.2.1
hmt-basemodels==0.0.22
hmt-escrow==0.8.11
httptools==0.1.1
idna==2.10
ifaddr==0.1.7
importlib-metadata==1.7.0
ipfshttpclient==0.6.1
ipython==7.9.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lahja==0.16.0
libp2p==0.1.5
lru-dict==1.1.6
lxml==4.5.2
MarkupSafe==1.1.1
marshmallow==3.7.1
multiaddr==0.0.9
multidict==4.7.6
mypy==0.782
mypy-extensions==0.4.3
netaddr==0.8.0
netdisco==2.8.2
netifaces==0.10.9
noiseprotocol==0.3.1
outcome==1.0.1
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plyvel==1.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
py-ecc==4.0.0
py-evm==0.3.0a19
py-solc-x==1.0.0
pycparser==2.20
pycryptodome==3.9.8
pydantic==1.6.1
pyethash==0.1.27
pyformance==0.4
Pygments==2.6.1
pymultihash==0.8.2
PyNaCl==1.3.0
pyrsistent==0.16.0
pysha3==1.0.2
python-dateutil==2.8.1
python-snappy==0.5.4
requests==2.24.0
rlp==2.0.0a1
rpcudp==3.0.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
rusty-rlp==0.1.15
s3transfer==0.3.3
schematics==2.1.0
semantic-version==2.8.5
six==1.15.0
sniffio==1.1.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.19
starlette==0.13.6
termcolor==1.1.0
toolz==0.10.0
traitlets==5.0.2
trie==2.0.0a4
trinity @ git+https://github.com/ethereum/trinity@988cff0dd55df75f6782346846fbd1f04cb908a7
trio==0.13.0
trio-typing==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
u-msgpack-python==2.7.0
uPnPClient==0.0.8
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
varint==1.0.2
wcwidth==0.2.5
web3==5.12.1
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.5.1
zeroconf==0.28.3
zipp==3.1.0
